backend/requirements.txt
fastapi
uvicorn
pydantic
networkx
openai
python-dotenv